pyproject.toml
[project]
name = "pokedex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.8",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.136.0",
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.0",
    "langchain-exa>=1.1.0",
    "langchain-groq>=1.1.2",
    "langchain-sarvam-integration>=0.1.9",
    "langchain-tavily>=0.2.18",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.56.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "uvicorn>=0.44.0",
]
